+++ src/main/java/com/usa/ciclo3/retociclo3/repository/ReservationRepository.java
@@ -48,7 +48,7 @@ public class ReservationRepository {
         List<CountClient> clientList = new ArrayList<>();
         List<Object[]> report = reservationCrudRepository.countTotalReservationByClient();
         for(int i=0;i<report.size();i++){
-            clientList.add(new CountClient((Double) report.get(i)[1],(Client)report.get(i)[0]));
+            clientList.add(new CountClient((Long) report.get(i)[1],(Client)report.get(i)[0]));
         }
         return clientList;
     }
